Move utility function.
