+++ src/lib/util/constructEventSlug.test.ts
