+++ src/lib/util/constructEventSlug.ts
